Optimize dag for dsl ci
